Add expect-argument-error.ts
