+++ packages/components/src/checkbox-group.test.basics.ts
@@ -2,6 +2,7 @@ import './checkbox.js';
 import { ArgumentError } from 'ow';
 import { elementUpdated, expect, fixture, html } from '@open-wc/testing';
 import CsCheckboxGroup from './checkbox-group.js';
+import expectArgumentError from './library/expect-argument-error.js';
 import sinon from 'sinon';
 
 CsCheckboxGroup.shadowRootOptions.mode = 'open';
@@ -48,7 +49,7 @@ it('can have a label', async () => {
   const component = await fixture<CsCheckboxGroup>(
     html`<cs-checkbox-group label="Checkbox Group">
       <cs-checkbox label="Checkbox"></cs-checkbox>
-    </cs-checkbox-group> `,
+    </cs-checkbox-group>`,
   );
 
   expect(component.getAttribute('label')).to.equal('Checkbox Group');
@@ -74,7 +75,7 @@ it('can have a name', async () => {
   const component = await fixture<CsCheckboxGroup>(
     html`<cs-checkbox-group label="Checkbox Group" name="name">
       <cs-checkbox label="Checkbox"></cs-checkbox>
-    </cs-checkbox-group> `,
+    </cs-checkbox-group>`,
   );
 
   expect(component.getAttribute('name')).to.equal('name');
@@ -101,7 +102,7 @@ it('can be disabled initially', async () => {
     html`<cs-checkbox-group label="Checkbox Group" disabled>
       <cs-checkbox label="One"></cs-checkbox>
       <cs-checkbox label="Two"></cs-checkbox>
-    </cs-checkbox-group> `,
+    </cs-checkbox-group>`,
   );
 
   const checkboxes = component.querySelectorAll('cs-checkbox');
@@ -118,7 +119,7 @@ it('can be disabled dynamically', async () => {
     html`<cs-checkbox-group label="Checkbox Group">
       <cs-checkbox label="One"></cs-checkbox>
       <cs-checkbox label="Two"></cs-checkbox>
-    </cs-checkbox-group> `,
+    </cs-checkbox-group>`,
   );
 
   component.disabled = true;
@@ -137,7 +138,7 @@ it('can be required', async () => {
   const component = await fixture<CsCheckboxGroup>(
     html`<cs-checkbox-group label="Checkbox Group" required>
       <cs-checkbox label="Checkbox"></cs-checkbox>
-    </cs-checkbox-group> `,
+    </cs-checkbox-group>`,
   );
 
   expect(component.hasAttribute('required')).to.be.true;
@@ -166,7 +167,7 @@ it('throws if it does not have a default slot', async () => {
 
   try {
     await fixture<CsCheckboxGroup>(
-      html`<cs-checkbox-group label="Checkbox Group"></cs-checkbox-group> `,
+      html`<cs-checkbox-group label="Checkbox Group"></cs-checkbox-group>`,
     );
   } catch (error) {
     if (error instanceof ArgumentError) {
@@ -178,33 +179,11 @@ it('throws if it does not have a default slot', async () => {
 });
 
 it('throws if the default slot is the incorrect type', async () => {
-  const onerror = window.onerror;
-
-  // Prevent Mocha from failing the test because of the failed "slotchange" assertion,
-  // which is expected. We'd catch the error below. But it happens in an event handler
-  // and so propagates to the window.
-  //
-  // https://github.com/mochajs/mocha/blob/99601da68d59572b6aa931e9416002bcb5b3e19d/browser-entry.js#L75
-  //
-  // eslint-disable-next-line unicorn/prefer-add-event-listener
-  window.onerror = null;
-
-  const spy = sinon.spy();
-
-  try {
-    await fixture<CsCheckboxGroup>(
+  await expectArgumentError(() => {
+    return fixture<CsCheckboxGroup>(
       html`<cs-checkbox-group label="Checkbox Group">
         <button>Button</button>
-      </cs-checkbox-group> `,
+      </cs-checkbox-group>`,
     );
-  } catch (error) {
-    if (error instanceof ArgumentError) {
-      spy();
-    }
-  }
-
-  expect(spy.called).to.be.true;
-
-  // eslint-disable-next-line unicorn/prefer-add-event-listener
-  window.onerror = onerror;
+  });
 });
