+++ packages/components/src/menu.test.basics.ts
@@ -3,6 +3,7 @@ import { ArgumentError } from 'ow';
 import { expect, fixture, html } from '@open-wc/testing';
 import CsMenu from './menu.js';
 import CsMenuLink from './menu.link.js';
+import expectArgumentError from './library/expect-argument-error.js';
 import sinon from 'sinon';
 
 CsMenu.shadowRootOptions.mode = 'open';
@@ -84,36 +85,14 @@ it('throws if it does not have a default slot', async () => {
 });
 
 it('throws if the default slot is the incorrect type', async () => {
-  const onerror = window.onerror;
-
-  // Prevent Mocha from failing the test because of the failed "slotchange" assertion,
-  // which is expected. We'd catch the error below. But it happens in an event handler
-  // and so propagates to the window.
-  //
-  // https://github.com/mochajs/mocha/blob/99601da68d59572b6aa931e9416002bcb5b3e19d/browser-entry.js#L75
-  //
-  // eslint-disable-next-line unicorn/prefer-add-event-listener
-  window.onerror = null;
-
-  const spy = sinon.spy();
-
-  try {
-    await fixture<CsMenu>(
+  await expectArgumentError(() => {
+    return fixture<CsMenu>(
       html`<cs-menu>
         <option>Option</option>
         <button slot="target">Target</button>
       </cs-menu>`,
     );
-  } catch (error) {
-    if (error instanceof ArgumentError) {
-      spy();
-    }
-  }
-
-  expect(spy.called).to.be.true;
-
-  // eslint-disable-next-line unicorn/prefer-add-event-listener
-  window.onerror = onerror;
+  });
 });
 
 it('throws if it does not have a "target" slot', async () => {
